requirements/dev.txt
robotframework
pytest
pylint
Pygments
docutils
